# gh-118761: Always lazy import `re` in `locale`