--- a/Lib/locale.py
+++ b/Lib/locale.py
@@ -176,6 +176,8 @@ def _strip_padding(s, amount):
         amount -= 1
     return s[lpos:rpos+1]
 
+_percent_re = None
+
 def _format(percent, value, grouping=False, monetary=False, *additional):
     if additional:
         formatted = percent % ((value,) + additional)
@@ -213,10 +215,12 @@ def format_string(f, val, grouping=False, monetary=False):
     Grouping is applied if the third parameter is true.
     Conversion uses monetary thousands separator and grouping strings if
     forth parameter monetary is true."""
-    import re
+    if _percent_re is None:
+        import re
 
-    _percent_re = re.compile(r'%(?:\((?P<key>.*?)\))?(?P<modifiers'
-                             r'>[-#0-9 +*.hlL]*?)[eEfFgGdiouxXcrs%]')
+        global _percent_re
+        _percent_re = re.compile(r'%(?:\((?P<key>.*?)\))?(?P<modifiers'
+                                 r'>[-#0-9 +*.hlL]*?)[eEfFgGdiouxXcrs%]')
 
     percents = list(_percent_re.finditer(f))
     new_f = _percent_re.sub('%s', f)