@picnixz 2025-02-08
But yes, that's what we wanted to see (namely tuna + hyperfine benchmarks). Ideally, I'd also like to have the benchmarks for `locale.format_string` now since the call is more complex.