@donBarbos 2025-02-08
I also found a weird argument passing for the format_string function in the same file on line 332:
https://github.com/python/cpython/blob/c1f352bf0813803bb795b796c16040a5cd4115f2/Lib/locale.py#L329-L336

I understand why it works, but can I change it to the expected argument type: `format_string("%d", 123456789, True)`?