@picnixz 2025-02-09
Sorry but that's not what I meant by "caches". The cache in question are just LRU cache at Python level:

https://github.com/python/cpython/blob/3cf859056a6e745cf056a9cb8b04e367bcbf5409/Lib/re/__init__.py#L330-L337

so it has nothing to do with `/proc/sys/vm/drop_caches`. In order to properly benchmark functions, we usually use `pyperf` instead. The idea is to essentially benchmark the call of `format_string`. I suspect that the first call to `format_string` will always be slower as we need to compile the pattern. Then, if we continue using `format_string` with any call to `re.compile()` in between, there won't be any issue. However, if we do multiple `re.compile()` with different patterns, then the next call to `format_string` will be slower as we need to make a full re-compile.