@donBarbos 2025-02-09
@picnixz 
> However, if we do multiple re.compile() with different patterns, then the next call to format_string will be slower as we need to make a full re-compile.

Now I understand, sorry :-)
But how can it happen that call do multiple `re.compile()` with different patterns if the pattern is hardcoded (the user does not pass it)