@picnixz 2025-02-09
> But how can it happen that call do multiple re.compile() with different patterns if the pattern is hardcoded (the user does not pass it)

Simple example:

```py
locale.format_string(...)
some_other_function_that_calls_re_compile()
...
locale.format_string(...)  # and we need to re-compile it
```

The `re.compile()` cache is shared not by "location" but by the entire Python process. So it might happen that you have other `re.compile()` calls between different calls to `locale.format_string()`.

Anyway, apart from this possibility of performance loss, the readability is still affected IMO. I'm not against Adam giving his opinion on this matter but if we want to have a real performance improvement, I think we should use the `_pattern = None` approach where you cache the compiled pattern at the module (`locale`) level:

```py
_pattern = None

def format_string(...):
    if _pattern is None:
        import re

        global _pattern
        _pattern = re.compile(...)
    ...
```    

With this approach `re.compile()` is called only once and for all but readability is affected even more.