@donBarbos 2025-02-09
Thank you for your explanation.
Then I will send a change that at least does not decrease performance. And I will leave this PR until the next reaction