@eli-schwartz 2025-02-09
> This will always recompile the pattern, which is not ideal. In addition, this function is likely to be used by `gettext` or other stuff so it will be called quite a lot IMO. Note that only the recent calls to `re.compile()` are cached, so it could or could not be efficient and it could also badly interact with other occurrences of `re.compile()`. An alternative is to use `_percent_re = None` + one-time compile but that could also harm readability.
> 
> I think I left out all `re` imports that were needed for a global pattern (if that pattern was necessary in a common function). Again, the issue is not just about making the import faster.
> 
> So, I'm not really convinced by this specific change. But maybe I'm overthinking and we can do it. Now, `locale` is likely to be used in text processing context where `re` is likely to have already been imported.

I did that None pattern in GH-128898, though admittedly my rationale for there was that it is quite "common" for gettext to be imported in contexts where it is installed as a global `_()` handler for, uh, readability purposes :) even if it never ends up being invoked -- and even if you turn out not to have any actual translations.

Note though that the import for this PR is only used by `locale.format_string()` which isn't used by gettext at all. In fact, nothing in the standard library uses this function at all. While there are surely non-stdlib uses of `locale.format_string()` it seems plausible that it isn't the default use case of the module :) so at the very least, avoiding a re.compile could be a bit of a time-saver even if your application still imports re either way...

It also seems like it should be relatively common to import locale to get at getencoding without necessarily using re or generalized text manipulation, just needing to know the correct format to open files or even some pesky subprocesses which a dreadful habit of communicating via non-UTF8 stdout.